removing old classes and renaming new classes
