+++ src/test/scala/com/ubirch/signatureverifier/RoutingTest.scala
@@ -1,19 +1,3 @@
-/*
- * Copyright (c) 2019 ubirch GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package com.ubirch.signatureverifier
 
 import java.util.{Base64, UUID}
@@ -23,124 +7,120 @@ import com.typesafe.scalalogging.StrictLogging
 import com.ubirch.client.protocol.DefaultProtocolVerifier
 import com.ubirch.client.util.curveFromString
 import com.ubirch.crypto.{GeneratorKeyFactory, PubKey}
-import com.ubirch.kafka.{EnvelopeDeserializer, EnvelopeSerializer, MessageEnvelope}
 import com.ubirch.niomon.base.{NioMicroservice, NioMicroserviceMock}
-import com.ubirch.protocol.ProtocolMessage
-import com.ubirch.protocol.codec.{JSONProtocolDecoder, MsgPackProtocolDecoder}
+import javax.xml.bind.DatatypeConverter
 import org.apache.kafka.clients.producer.ProducerRecord
+import org.apache.kafka.common.header.internals.{RecordHeader, RecordHeaders}
+import org.apache.kafka.common.serialization.{ByteArrayDeserializer, ByteArraySerializer}
 import org.scalatest.{FlatSpec, Matchers}
 
-//noinspection TypeAnnotation
 class RoutingTest extends FlatSpec with Matchers with StrictLogging {
-  implicit val messageEnvelopeSerializer = EnvelopeSerializer
-  implicit val messageEnvelopeDeserializer = EnvelopeDeserializer
 
-  val keyServerClient = (c: NioMicroservice.Context) => new CachingUbirchKeyService(c) {
-    val eddsaKey = GeneratorKeyFactory.getPubKey(
-      Base64.getDecoder.decode("sSqQYFHxAogbu0h+6CZKoF2ND8xRIY8qR/Vizrmw0Gg="),
-      curveFromString("ECC_ED25519")
-    )
+  implicit val bytesSerializer: ByteArraySerializer = new ByteArraySerializer
+  implicit val byteDeserializer: ByteArrayDeserializer = new ByteArrayDeserializer
 
-    val waldisKey = GeneratorKeyFactory.getPubKey(
-      Base64.getDecoder.decode("5hwcViaiHqnKRly7aFSqWEPH75TfFn7FydqxJVDrQIY="),
-      curveFromString("ECC_ED25519")
-    )
+  private val v2HardwareId = "6eac4d0b-16e6-4508-8c46-22e7451ea5a1"
+  private val v2MsgPackHex = "9522c4106eac4d0b16e645088c4622e7451ea5a100c4206b86b273ff34fce19d6b804eff5a3f5747ada4eaa22f1d49c01e52ddb7875b4bc440bc2a01322c679b9648a9391704e992c041053404aafcdab08fc4ce54a57eb16876d741918d01219abf2dc7913f2d9d49439d350f11d05cdb3f85972ac95c45fc"
 
-    val ecdsaKey = GeneratorKeyFactory.getPubKey(
-      Base64.getDecoder.decode("kvdvWQ7NOT+HLDcrFqP/UZWy4QVcjfmmkfyzAgg8bitaK/FbHUPeqEji0UmCSlyPk5+4mEaEiZAHnJKOyqUZxA=="),
+  private val v1HardwareId = "946ba755-e132-4680-a380-7be395eb8ccd"
+  private val v1MsgPackHex = "9613b0946ba755e1324680a3807be395eb8ccdda004001d3c78a7bc8738cbabee96ae86651c89aa81252dee0e7c7f3457324202aae9c3d16332d3d7ecde72fa26801a081b6347148bae6ab22b8830c360d78d0b63f0e00a50102030405da0040b30fa8a898282878862e15ed871f6618043fb7c2400daed1bdf7bad5c790edd5b011d248be2ee04bc086b02bd9c337b981eeaff1eefca47d346132ab4cfd8006"
+
+  private val trackleHardwareId = "7aa42ea9-f981-91a6-9b9c-b14aacabce54"
+  private val trackleMsgPack = "96cd0013b07aa42ea9f98191a69b9cb14aacabce54da0040863af1a1ed5d1051566ab79897347b36b86cf2553efb8509a13e440276e3327653f9d02a09ab2904803dfd6b632972f110b32090f3a706592054c5ad792ce1025495da002376312e302e322d50524f442d3230313830333236313033323035202876352e362e3629cd15aa03de01edce5ebb05ebcd0e6ace5ebb0627cd0e6ace5ebb0663cd0e6ace5ebb069fcd0e6ace5ebb06dbcd0e69ce5ebb0717cd0e69ce5ebb0753cd0e68ce5ebb078fcd0e69ce5ebb07cbcd0e67ce5ebb0807cd0e65ce5ebb0843cd0e64ce5ebb087fcd0e64ce5ebb08bbcd0e64ce5ebb08f7cd0e61ce5ebb0933cd0e5fce5ebb096fcd0e5fce5ebb09abcd0e5dce5ebb09e7cd0e5ece5ebb0a23cd0e5ece5ebb0a5fcd0e5cce5ebb0a9bcd0e5bce5ebb0ad7cd0e59ce5ebb0b13cd0e57ce5ebb0b4fcd0e55ce5ebb0b8bcd0e54ce5ebb0bc7cd0e55ce5ebb0c03cd0e53ce5ebb0c3fcd0e55ce5ebb0c7bcd0e56ce5ebb0cb7cd0e54ce5ebb0cf3cd0e52ce5ebb0d2fcd0e52ce5ebb0d6bcd0e52ce5ebb0da7cd0e55ce5ebb0de3cd0e56ce5ebb0e1fcd0e54ce5ebb0e5bcd0e54ce5ebb0e97cd0e54ce5ebb0ed3cd0e54ce5ebb0f0fcd0e56ce5ebb0f4bcd0e54ce5ebb0f87cd0e56ce5ebb0fc3cd0e56ce5ebb0fffcd0e57ce5ebb103bcd0e57ce5ebb1077cd0e59ce5ebb10b3cd0e57ce5ebb10efcd0e55ce5ebb112bcd0e57ce5ebb1167cd0e56ce5ebb11a3cd0e57ce5ebb11dfcd0e55ce5ebb121bcd0e56ce5ebb1257cd0e55ce5ebb1293cd0e56ce5ebb12cfcd0e54ce5ebb130bcd0e52ce5ebb1347cd0e52ce5ebb1383cd0e52ce5ebb13bfcd0e52ce5ebb13fbcd0e54ce5ebb1437cd0e54ce5ebb1473cd0e54ce5ebb14afcd0e54ce5ebb14ebcd0e53ce5ebb1527cd0e54ce5ebb1563cd0e52ce5ebb159fcd0e52ce5ebb15dbcd0e51ce5ebb1617cd0e54ce5ebb1653cd0e52ce5ebb168fcd0e52ce5ebb16cbcd0e52ce5ebb1707cd0e52ce5ebb1743cd0e51ce5ebb177fcd0e52ce5ebb17bbcd0e51ce5ebb17f7cd0e52ce5ebb1833cd0e53ce5ebb186fcd0e55ce5ebb18abcd0e52ce5ebb18e7cd0e54ce5ebb1923cd0e53ce5ebb195fcd0e52ce5ebb199bcd0e53ce5ebb19d7cd0e54ce5ebb1a13cd0e55ce5ebb1a4fcd0e54ce5ebb1a8bcd0e54ce5ebb1ac7cd0e53ce5ebb1b03cd0e55ce5ebb1b3fcd0e53ce5ebb1b7bcd0e54ce5ebb1bb7cd0e53ce5ebb1bf3cd0e56ce5ebb1c2fcd0e53ce5ebb1c6bcd0e55ce5ebb1ca7cd0e56ce5ebb1ce3cd0e59ce5ebb1d1fcd0e59ce5ebb1d5bcd0e57ce5ebb1d97cd0e56ce5ebb1dd3cd0e54ce5ebb1e0fcd0e57ce5ebb1e4bcd0e59ce5ebb1e87cd0e56ce5ebb1ec3cd0e56ce5ebb1effcd0e56ce5ebb1f3bcd0e59ce5ebb1f77cd0e56ce5ebb1fb3cd0e59ce5ebb1fefcd0e59ce5ebb202bcd0e57ce5ebb2067cd0e5ace5ebb20a3cd0e5bce5ebb20dfcd0e5ace5ebb211bcd0e5ace5ebb2157cd0e5cce5ebb2193cd0e5bce5ebb21cfcd0e5bce5ebb220bcd0e5bce5ebb2247cd0e5dce5ebb2283cd0e5dce5ebb22bfcd0e5cce5ebb22fbcd0e5bce5ebb2337cd0e59ce5ebb2373cd0e5bce5ebb23afcd0e5cce5ebb23ebcd0e5cce5ebb2427cd0e5cce5ebb2463cd0e5dce5ebb249fcd0e5dce5ebb24dbcd0e5dce5ebb2517cd0e5dce5ebb2553cd0e5ece5ebb258fcd0e60ce5ebb25cbcd0e5fce5ebb2607cd0e5fce5ebb2643cd0e5fce5ebb267fcd0e5ece5ebb26bbcd0e5ece5ebb26f7cd0e5dce5ebb2733cd0e5fce5ebb276fcd0e61ce5ebb27abcd0e61ce5ebb27e7cd0e5ece5ebb2823cd0e5fce5ebb285fcd0e61ce5ebb289bcd0e62ce5ebb28d7cd0e62ce5ebb2913cd0e60ce5ebb294fcd0e62ce5ebb298bcd0e62ce5ebb29c7cd0e62ce5ebb2a03cd0e61ce5ebb2a3fcd0e62ce5ebb2a7bcd0e63ce5ebb2ab7cd0e62ce5ebb2af3cd0e63ce5ebb2b2fcd0e64ce5ebb2b6bcd0e62ce5ebb2ba7cd0e64ce5ebb2be3cd0e64ce5ebb2c1fcd0e63ce5ebb2c5bcd0e64ce5ebb2c97cd0e67ce5ebb2cd3cd0e65ce5ebb2d0fcd0e64ce5ebb2d4bcd0e68ce5ebb2d87cd0e68ce5ebb2dc3cd0e67ce5ebb2dffcd0e68ce5ebb2e3bcd0e68ce5ebb2e77cd0e69ce5ebb2eb3cd0e69ce5ebb2eefcd0e69ce5ebb2f2bcd0e69ce5ebb2f67cd0e6cce5ebb2fa3cd0e6bce5ebb2fdfcd0e6bce5ebb301bcd0e6cce5ebb3057cd0e6dce5ebb3093cd0e6cce5ebb30cfcd0e6dce5ebb310bcd0e6ece5ebb3147cd0e6fce5ebb3183cd0e6ece5ebb31bfcd0e6fce5ebb31fbcd0e6dce5ebb3237cd0e6fce5ebb3273cd0e6fce5ebb32afcd0e6fce5ebb32ebcd0e6fce5ebb3327cd0e6ece5ebb3363cd0e6fce5ebb339fcd0e6fce5ebb33dbcd0e6ece5ebb3417cd0e6fce5ebb3453cd0e6dce5ebb348fcd0e6ece5ebb34cbcd0e6fce5ebb3507cd0e6fce5ebb3543cd0e6fce5ebb357fcd0e72ce5ebb35bbcd0e72ce5ebb35f7cd0e70ce5ebb3633cd0e6fce5ebb366fcd0e70ce5ebb36abcd0e70ce5ebb36e7cd0e71ce5ebb3723cd0e71ce5ebb375fcd0e71ce5ebb379bcd0e72ce5ebb37d7cd0e74ce5ebb3813cd0e71ce5ebb384fcd0e73ce5ebb388bcd0e74ce5ebb38c7cd0e74ce5ebb3903cd0e73ce5ebb393fcd0e72ce5ebb397bcd0e70ce5ebb39b7cd0e72ce5ebb39f3cd0e72ce5ebb3a2fcd0e71ce5ebb3a6bcd0e74ce5ebb3aa7cd0e6fce5ebb3ae3cd0e72ce5ebb3b1fcd0e70ce5ebb3b5bcd0e71ce5ebb3b97cd0e71ce5ebb3bd3cd0e70ce5ebb3c0fcd0e6fce5ebb3c4bcd0e6fce5ebb3c87cd0e6ece5ebb3cc3cd0e6cce5ebb3cffcd0e6cce5ebb3d3bcd0e6cce5ebb3d77cd0e6ece5ebb3db3cd0e6bce5ebb3defcd0e6ace5ebb3e2bcd0e6dce5ebb3e67cd0e69ce5ebb3ea3cd0e69ce5ebb3edfcd0e6ace5ebb3f1bcd0e69ce5ebb3f57cd0e6ace5ebb3f93cd0e67ce5ebb3fcfcd0e67ce5ebb400bcd0e67ce5ebb4047cd0e63ce5ebb4083cd0e64ce5ebb40bfcd0e62ce5ebb40fbcd0e62ce5ebb4137cd0e63ce5ebb4173cd0e62ce5ebb41afcd0e62ce5ebb41ebcd0e61ce5ebb4227cd0e5ece5ebb4263cd0e5fce5ebb429fcd0e5fce5ebb42dbcd0e60ce5ebb4317cd0e5ece5ebb4353cd0e5dce5ebb438fcd0e5cce5ebb43cbcd0e5bce5ebb4407cd0e5bce5ebb4443cd0e5ace5ebb447fcd0e5bce5ebb44bbcd0e5cce5ebb44f7cd0e5cce5ebb4533cd0e5dce5ebb456fcd0e60ce5ebb45abcd0e5dce5ebb45e7cd0e5ece5ebb4623cd0e60ce5ebb465fcd0e5ece5ebb469bcd0e5dce5ebb46d7cd0e5cce5ebb4713cd0e5fce5ebb474fcd0e5fce5ebb478bcd0e5ece5ebb47c7cd0e5ece5ebb4803cd0e5fce5ebb483fcd0e60ce5ebb487bcd0e5fce5ebb48b7cd0e5ece5ebb48f3cd0e5ece5ebb492fcd0e5ece5ebb496bcd0e5ece5ebb49a7cd0e60ce5ebb49e3cd0e60ce5ebb4a1fcd0e5ece5ebb4a5bcd0e5fce5ebb4a97cd0e5ece5ebb4ad3cd0e5ece5ebb4b0fcd0e5ece5ebb4b4bcd0e5ece5ebb4b87cd0e5dce5ebb4bc3cd0e5ece5ebb4bffcd0e5fce5ebb4c3bcd0e5ece5ebb4c77cd0e5dce5ebb4cb3cd0e5ece5ebb4cefcd0e5dce5ebb4d2bcd0e60ce5ebb4d67cd0e5dce5ebb4da3cd0e5ece5ebb4ddfcd0e60ce5ebb4e1bcd0e5ece5ebb4e57cd0e60ce5ebb4e93cd0e5dce5ebb4ecfcd0e5ece5ebb4f0bcd0e5ece5ebb4f47cd0e5fce5ebb4f83cd0e5ece5ebb4fbfcd0e5fce5ebb4ffccd0e5ece5ebb5038cd0e5ece5ebb5074cd0e5ece5ebb50b0cd0e5dce5ebb50eccd0e5dce5ebb5128cd0e5ece5ebb5164cd0e5dce5ebb51a0cd0e5ece5ebb51dccd0e5cce5ebb5218cd0e5ece5ebb5254cd0e5ece5ebb5290cd0e5dce5ebb52cccd0e5dce5ebb5308cd0e5ece5ebb5344cd0e5dce5ebb5380cd0e5ece5ebb53bccd0e5ece5ebb53f8cd0e5ece5ebb5434cd0e5ece5ebb5470cd0e5dce5ebb54accd0e5ece5ebb54e8cd0e5ece5ebb5524cd0e5dce5ebb5560cd0e5ece5ebb559ccd0e5dce5ebb55d8cd0e5ece5ebb5614cd0e5dce5ebb5650cd0e5dce5ebb568ccd0e5fce5ebb56c8cd0e5ece5ebb5704cd0e5ece5ebb5740cd0e5cce5ebb577ccd0e5dce5ebb57b8cd0e5ece5ebb57f4cd0e5ece5ebb5830cd0e5cce5ebb586ccd0e5dce5ebb58a8cd0e5cce5ebb58e4cd0e5dce5ebb5920cd0e5dce5ebb595ccd0e5ece5ebb5998cd0e5fce5ebb59d4cd0e60ce5ebb5a10cd0e5dce5ebb5a4ccd0e5bce5ebb5a88cd0e5dce5ebb5ac4cd0e5cce5ebb5b00cd0e5dce5ebb5b3ccd0e5ece5ebb5b78cd0e5ece5ebb5bb4cd0e5dce5ebb5bf0cd0e5dce5ebb5c2ccd0e5ece5ebb5c68cd0e5dce5ebb5ca4cd0e5dce5ebb5ce0cd0e5ece5ebb5d1ccd0e5fce5ebb5d58cd0e61ce5ebb5d94cd0e60ce5ebb5dd0cd0e61ce5ebb5e0ccd0e5ece5ebb5e48cd0e5fce5ebb5e84cd0e5ece5ebb5ec0cd0e5dce5ebb5efccd0e5fce5ebb5f38cd0e5dce5ebb5f74cd0e5ece5ebb5fb0cd0e5ece5ebb5feccd0e60ce5ebb6028cd0e5fce5ebb6064cd0e61ce5ebb60a0cd0e61ce5ebb60dccd0e62ce5ebb6118cd0e62ce5ebb6154cd0e62ce5ebb6190cd0e64ce5ebb61cccd0e62ce5ebb6208cd0e64ce5ebb6244cd0e65ce5ebb6280cd0e65ce5ebb62bccd0e65ce5ebb62f8cd0e65ce5ebb6334cd0e67ce5ebb6370cd0e63ce5ebb63accd0e65ce5ebb63e8cd0e63ce5ebb6424cd0e64ce5ebb6460cd0e63ce5ebb649ccd0e64ce5ebb64d8cd0e64ce5ebb6514cd0e69ce5ebb6550cd0e65ce5ebb658ccd0e67ce5ebb65c8cd0e65ce5ebb6604cd0e67ce5ebb6640cd0e67ce5ebb667ccd0e67ce5ebb66b8cd0e67ce5ebb66f4cd0e67ce5ebb6730cd0e67ce5ebb676ccd0e67ce5ebb67a8cd0e64ce5ebb67e4cd0e64ce5ebb6820cd0e67ce5ebb685ccd0e64ce5ebb6898cd0e64ce5ebb68d4cd0e67ce5ebb6910cd0e65ce5ebb694ccd0e67ce5ebb6988cd0e68ce5ebb69c4cd0e68ce5ebb6a00cd0e67ce5ebb6a3ccd0e69ce5ebb6a78cd0e68ce5ebb6ab4cd0e67ce5ebb6af0cd0e68ce5ebb6b2ccd0e68ce5ebb6b68cd0e67ce5ebb6ba4cd0e69ce5ebb6be0cd0e69ce5ebb6c1ccd0e69ce5ebb6c58cd0e67ce5ebb6c94cd0e6ace5ebb6cd0cd0e6ace5ebb6d0ccd0e68ce5ebb6d48cd0e6ace5ebb6d84cd0e6ace5ebb6dc0cd0e68ce5ebb6dfccd0e6ace5ebb6e38cd0e69ce5ebb6e74cd0e6cce5ebb6eb0cd0e69ce5ebb6eeccd0e6bce5ebb6f28cd0e6bce5ebb6f64cd0e69ce5ebb6fa0cd0e6ace5ebb6fdccd0e69ce5ebb7018cd0e6ace5ebb7054cd0e6ace5ebb7090cd0e6bce5ebb70cccd0e6cce5ebb7108cd0e6cce5ebb7144cd0e6ace5ebb7180cd0e6cce5ebb71bccd0e6ece5ebb71f8cd0e6cce5ebb7234cd0e6dce5ebb7270cd0e6ece5ebb72accd0e6fce5ebb72e8cd0e6fce5ebb7324cd0e6cce5ebb7360cd0e6dce5ebb739ccd0e6cce5ebb73d8cd0e6ece5ebb7414cd0e6fce5ebb7450cd0e6fce5ebb748ccd0e6fce5ebb74c8cd0e6fce5ebb7504cd0e6fce5ebb7540cd0e6fce5ebb757ccd0e70ce5ebb75b8cd0e6fce5ebb75f4cd0e70ce5ebb7630cd0e70ce5ebb766ccd0e71ce5ebb76a8cd0e72ce5ebb76e4cd0e72ce5ebb7720cd0e71ce5ebb775ccd0e71ce5ebb7798cd0e71ce5ebb77d4cd0e72ce5ebb7810cd0e72ce5ebb784ccd0e71ce5ebb7888cd0e70ce5ebb78c4cd0e71ce5ebb7900cd0e70ce5ebb793ccd0dc584a36d696ecd0daca36d6178cd1068a169cdea60a2696cce001b7740da004046df900da3a6d847fcdd4b527628b6f13eb056707c633e5f6e76d14e873bf2f3f595f29342eb61abf91f45fd42ba55ac1098736ecfa6fe1c942f77091eb2680b"
+
+  private val keyServerClient = (c: NioMicroservice.Context) => new CachingUbirchKeyService(c) {
+    private val eddsaKey = GeneratorKeyFactory.getPubKey(
+      "55f0feac4f2bcf879330eff348422ab3abf5237a24acaf0aef3bb876045c4e532fbd6cd8e265f6cf28b46e7e4512cd06ba84bcd3300efdadf28750f43dafd771",
       curveFromString("ecdsa-p256v1")
+    )
+
+    private val waldisKey = GeneratorKeyFactory.getPubKey(
+      Base64.getDecoder.decode("7odnHwchQA6AAk8AZzCQevnWM2PT4Jg+MSxVDf7wGsU="),
+      curveFromString("ECC_ED25519")
+    )
+
+    private val trackleKey = GeneratorKeyFactory.getPubKey(
+      Base64.getDecoder.decode("Lujc3SSchaqzGn/wDZWzmRfwVgdCZBoW6QkLT7N1700="),
+      curveFromString("ECC_ED25519")
     )
 
     override def getPublicKey(uuid: UUID): List[PubKey] = {
       uuid.toString match {
-        case "6eac4d0b-16e6-4508-8c46-22e7451ea5a1" => List(eddsaKey)
-        case "ffff160c-6117-5b89-ac98-15aeb52655e0" => List(ecdsaKey)
-        case "30aea484-3b2c-1223-3445-566778899aab" => List(waldisKey)
+        case id if id == v2HardwareId => List(eddsaKey)
+        case id if id == v1HardwareId => List(waldisKey)
+        case id if id == trackleHardwareId => List(trackleKey)
         case _ => Nil
       }
     }
   }
 
-  val microservice = NioMicroserviceMock(SignatureVerifierMicroservice(c => new DefaultProtocolVerifier(keyServerClient(c))))
+  private val microservice = NioMicroserviceMock(SignatureVerifierMicroservice(c => new DefaultProtocolVerifier(keyServerClient(c))))
   microservice.outputTopics = Map("valid" -> "valid")
   microservice.errorTopic = Some("invalid")
   microservice.config = ConfigFactory.load().getConfig("niomon-verifier")
   microservice.name = "niomon-verifier"
+
   import microservice.kafkaMocks._
 
 
-  "msgpack with valid signature" should "be routed to 'valid' queue" in {
-    val message = Base64.getDecoder.decode("lRKwbqxNCxbmRQiMRiLnRR6loczvAdoAQFeKWyLOs+HQ0PiUfAmAEBM7RNOx0qs5h1j/7RFQe2B+03274Ab2RfDtD9vrG0i7UP1x2DI0DOAk1aDiHA68jg4=")
-    val pm = MsgPackProtocolDecoder.getDecoder.decode(message)
-    val validMessage = MessageEnvelope(pm)
-    logger.info(validMessage.toString)
+  "msgpackv2 with valid signature" should "be routed to 'valid' queue" in {
+    val binary = DatatypeConverter.parseHexBinary(v2MsgPackHex)
+    val record = new ProducerRecord[String, Array[Byte]]("incoming", 1, "foo", binary, new RecordHeaders().add(new RecordHeader("X-Ubirch-Hardware-Id".toLowerCase, v2HardwareId.getBytes())))
+    publishToKafka(record)
 
-    publishToKafka(new ProducerRecord("incoming", "foo", validMessage))
-
-    val validTopicEnvelopes = consumeNumberMessagesFrom[MessageEnvelope]("valid", 1, autoCommit = true)
+    val validTopicEnvelopes = consumeNumberMessagesFrom("valid", 1, autoCommit = true)
     validTopicEnvelopes.size should be(1)
 
     val approvedMessage = validTopicEnvelopes.head
-    approvedMessage.toString should equal(validMessage.toString) // ProtocolMessage doesn't override equals :'(
+    approvedMessage should equal(binary)
   }
 
-  "json with valid signature" should "be routed to 'valid' queue" in {
-    val message = "{\"version\":18,\"uuid\":\"6eac4d0b-16e6-4508-8c46-22e7451ea5a1\",\"hint\":239,\"signature\":\"YyC6ChlzkEOxL0oH98ytZz4ZOUEmE3uFlt3Ildy2X1/Pdp9BtSQvMScZKjUK6Y0berKHKR7LRYAwD7Ko+BBXCA==\",\"payload\":1}"
-    val pm = JSONProtocolDecoder.getDecoder.decode(message)
-    val validMessage = MessageEnvelope(pm)
-    logger.info(validMessage.toString)
+  "msgpackv1 with valid signature" should "be routed to 'valid' queue" in {
+    val binary = DatatypeConverter.parseHexBinary(v1MsgPackHex)
+    val record = new ProducerRecord[String, Array[Byte]]("incoming", 1, "foo", binary, new RecordHeaders().add(new RecordHeader("X-Ubirch-Hardware-Id".toLowerCase, v1HardwareId.getBytes())))
+    publishToKafka(record)
 
-    publishToKafka(new ProducerRecord("incoming", "foo", validMessage))
-
-    val validTopicEnvelopes = consumeNumberMessagesFrom[MessageEnvelope]("valid", 1, autoCommit = true)
+    val validTopicEnvelopes = consumeNumberMessagesFrom("valid", 1, autoCommit = true)
     validTopicEnvelopes.size should be(1)
 
     val approvedMessage = validTopicEnvelopes.head
-    approvedMessage.toString should equal(validMessage.toString) // ProtocolMessage doesn't override equals :'(
+    approvedMessage should equal(binary)
   }
 
-  "json with invalid signature " should "be routed to 'invalid' queue" in {
-    val invalidMessage = MessageEnvelope(new ProtocolMessage())
-    publishToKafka(new ProducerRecord("incoming", "bar", invalidMessage))
+  "msgpackv1 with invalid signature" should "be routed to 'invalid' queue" in {
+    val binary = DatatypeConverter.parseHexBinary(v1MsgPackHex + "12")
+    val record = new ProducerRecord[String, Array[Byte]]("incoming", 1, "foo", binary, new RecordHeaders().add(new RecordHeader("X-Ubirch-Hardware-Id".toLowerCase, v1HardwareId.getBytes())))
+    publishToKafka(record)
 
     val invalidTopicEnvelopes = consumeNumberStringMessagesFrom("invalid", 1, autoCommit = true)
     invalidTopicEnvelopes.size should be(1)
 
     val rejectedMessage = invalidTopicEnvelopes.head
-    rejectedMessage.toString should equal("""{"error":"NullPointerException: null","causes":[],"microservice":"niomon-verifier","requestId":"bar"}""")
+    rejectedMessage should equal("""{"error":"SignatureException: Invalid signature","causes":[],"microservice":"niomon-verifier","requestId":"foo"}""")
   }
 
-  "ecdsa verification" should "work" in {
-    val message = Base64.getDecoder.decode("lSLEEP//FgxhF1uJrJgVrrUmVeAAxECUnW4kkga5FhldAMYFX7s8ZUTQwYZpV3ObvNKa27c+wVoGfmGN9zQwPbl2hXBq2femGe6NzSjUtQwAIVMXrERexEBKdNrNNjCpzGR/PwNNxxIwjFL++EEoSquEAyW/JW5cPblVnxC+rIgt4+0gUFbWy5IAZcOmmvtDFeP/u/G1lIU7")
-    val pm = MsgPackProtocolDecoder.getDecoder.decode(message)
+  "msgpackv1 with no hwDeviceId in the header" should "be routed to 'valid' queue" in {
+    val binary = DatatypeConverter.parseHexBinary(v1MsgPackHex + "12")
+    val record = new ProducerRecord[String, Array[Byte]]("incoming", 1, "foo", binary, new RecordHeaders().add(new RecordHeader("No-Hardware-Id".toLowerCase, v1HardwareId.getBytes())))
+    publishToKafka(record)
 
-    logger.debug(Base64.getEncoder.encodeToString(pm.getSigned))
-    logger.debug(Base64.getEncoder.encodeToString(pm.getSignature))
+    val invalidTopicEnvelopes = consumeNumberMessagesFrom("invalid", 1, autoCommit = true)
+    invalidTopicEnvelopes.size should be(1)
 
-    val validMessage = MessageEnvelope(pm)
-    publishToKafka(new ProducerRecord("incoming", "foo", validMessage))
+    val rejectedMessage: Array[Byte] = invalidTopicEnvelopes.head
+    rejectedMessage.map(_.toChar).mkString should equal("{\"error\":\"SignatureException: Header with key x-ubirch-hardware-id is missing. Cannot verify msgPack.\",\"causes\":[],\"microservice\":\"niomon-verifier\",\"requestId\":\"foo\"}")
+  }
 
-    val validTopicEnvelopes = consumeNumberMessagesFrom[MessageEnvelope]("valid", 1, autoCommit = true)
+  "trackleMsg with valid signature" should "be routed to 'valid' queue" in {
+    val binary = DatatypeConverter.parseHexBinary(trackleMsgPack)
+    val record = new ProducerRecord[String, Array[Byte]]("incoming", 1, "foo", binary, new RecordHeaders().add(new RecordHeader("X-Ubirch-Hardware-Id".toLowerCase, trackleHardwareId.getBytes())))
+    publishToKafka(record)
+
+    val validTopicEnvelopes = consumeNumberMessagesFrom("valid", 1, autoCommit = true)
     validTopicEnvelopes.size should be(1)
 
     val approvedMessage = validTopicEnvelopes.head
-    approvedMessage.toString should equal(validMessage.toString) // ProtocolMessage doesn't override equals :'(
+    approvedMessage should equal(binary)
   }
 
-  "waldi's case" should "work" in {
-    val message = Base64.getDecoder.decode("ls0AE7AwrqSEOywSIzRFVmd4iZqr2gBAOIfVH/Ann3VM74vP4B1KHhfjLNHdJbbDUZYr+S30e3eBhFLTUiBybBODe+gzei0zAJ/YydtT9s7LldqT3bwlBQDaAEDcSagOlzSV94Brx434qiDmK/CLPyt5Ghg34HOrgeUo8TGQpIrhEp6bOdrDhTLiHsxD9+1OldhwKxFKdnZ0ZGdw2gBAA709I+fcOrrL2W7ZeICz6dl4x5ci7sx19wHE1so9CcMVa7wzY6BX2AyqPnE0LPs8NMYvJQA68E4m/n2tAU9tCQ==")
-    val pm = MsgPackProtocolDecoder.getDecoder.decode(message)
-
-    logger.debug(Base64.getEncoder.encodeToString(pm.getSigned))
-    logger.debug(Base64.getEncoder.encodeToString(pm.getSignature))
-
-    val validMessage = MessageEnvelope(pm)
-    publishToKafka(new ProducerRecord("incoming", "foo", validMessage))
-
-    val validTopicEnvelopes = consumeNumberMessagesFrom[MessageEnvelope]("valid", 1, autoCommit = true)
-    validTopicEnvelopes.size should be(1)
-
-    val approvedMessage = validTopicEnvelopes.head
-    approvedMessage.toString should equal(validMessage.toString) // ProtocolMessage doesn't override equals :'(
-  }
 }
